SEQUOIADBMAINSTREAM-1739 LOB性能优化，增加LOB缓存功能
git-svn-id: http://192.168.20.11/sequoiadb/trunk@23427 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdSession.cpp
@@ -237,7 +237,7 @@ namespace engine
                                             INT32 &buffLen )
    {
       INT32 rc = SDB_OK ;
-      INT32 recvSize = (INT32)sizeof(MsgSysInfoRequest) ;
+      UINT32 recvSize = sizeof(MsgSysInfoRequest) ;
 
       *ppBuff = getBuff( recvSize ) ;
       if ( !*ppBuff )
